+++ requestList/js/events.js
@@ -1,7 +1,5 @@
 //#region EVENTS
 function bindEvents() {
-  bindChangeRequestEvents();
-
   $(document).on("input", "#search-bar", function () {
     searchEmployee();
   });