+++ requestList/js/state.js
@@ -47,7 +47,5 @@ let presID = [];
 let reqAccess = false;
 let headerData = null;
 let selectedDispatchRequest = null;
-let changeRequestTriggerElement = null;
-let isChangeRequestSubmitting = false;
 const { jsPDF } = globalThis.jspdf;
 //#endregion